Calculate variable importance from ADM Datamart data.

Uses the predictor binning from the ADM Datamart data to find the relative importance. The importance is given by the distance to the average log odds, per predictor. Next to importance it also returns the univariate predictor performance (AUC).

admVarImp(
  datamart,
  facets = NULL,
  filter = function(x) {     filterClassifierOnly(x, reverse = T) }
)

Arguments

datamart
Data frame with the ADM datamart data. Model data is only used when there are facets, so the model part could in principle be NULL.

facets
Optional list of names from the model data to aggregate up to.

filter
Optional filter for the predictor data. Defaults to filter out the classifiers, so operates on both active and inactive predictors. See examples.

Value

A data.table) with PredictorName, Importance and Rank plus columns for each of the facets if supplied.

Examples

if (FALSE) {
  dm <- ADMDatamart("~/Downloads")

  varimp <- admVarImp(dm)

  varimp <- admVarImp(dm, filter = filterActiveOnly)
}