Subset the datamart predictor data to only the active predictors.

Subset the datamart predictor data to only the active predictors.

filterActiveOnly(dt, reverse = F)

Arguments

dt

Predictor datamart table

reverse

Boolean to flip the operation.

Value

A data.table.

Examples

data(adm_datamart)
filterActiveOnly(adm_datamart$predictordata)
#>                                              PredictorName
#>     1:               IH.SMS.Outbound.Churned.pxLastGroupID
#>     2:               IH.SMS.Outbound.Churned.pxLastGroupID
#>     3:               IH.SMS.Outbound.Churned.pxLastGroupID
#>     4:               IH.SMS.Outbound.Churned.pxLastGroupID
#>     5:               IH.SMS.Outbound.Churned.pxLastGroupID
#>    ---                                                    
#> 30367: IH.SMS.Outbound.Churned.pxLastOutcomeTime.DaysSince
#> 30368: IH.SMS.Outbound.Churned.pxLastOutcomeTime.DaysSince
#> 30369: IH.SMS.Outbound.Churned.pxLastOutcomeTime.DaysSince
#> 30370: IH.SMS.Outbound.Churned.pxLastOutcomeTime.DaysSince
#> 30371: IH.SMS.Outbound.Churned.pxLastOutcomeTime.DaysSince
#>                                     ModelID     Type      BinType
#>     1: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7 symbolic      MISSING
#>     2: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7 symbolic      MISSING
#>     3: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7 symbolic EQUIBEHAVIOR
#>     4: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7 symbolic EQUIBEHAVIOR
#>     5: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7 symbolic      MISSING
#>    ---                                                           
#> 30367: c3f445b2-6037-5560-9f16-71a7259d2c2c  numeric EQUIBEHAVIOR
#> 30368: c3f445b2-6037-5560-9f16-71a7259d2c2c  numeric EQUIBEHAVIOR
#> 30369: c3f445b2-6037-5560-9f16-71a7259d2c2c  numeric EQUIBEHAVIOR
#> 30370: c3f445b2-6037-5560-9f16-71a7259d2c2c  numeric EQUIBEHAVIOR
#> 30371: c3f445b2-6037-5560-9f16-71a7259d2c2c  numeric EQUIBEHAVIOR
#>                BinSymbol BinLowerBound BinUpperBound EntryType
#>     1:           MISSING            NA            NA    Active
#>     2:           MISSING            NA            NA    Active
#>     3:       NON-MISSING            NA            NA    Active
#>     4:       NON-MISSING            NA            NA    Active
#>     5:           MISSING            NA            NA    Active
#>    ---                                                        
#> 30367:           <0.0831   -112.876946      0.083054    Active
#> 30368: [0.0831, 16.4031>      0.083054     16.403054    Active
#> 30369: [0.0831, 16.4031>      0.083054     16.403054    Active
#> 30370:         >=16.4031     16.403054    143.123054    Active
#> 30371:         >=16.4031     16.403054    143.123054    Active
#>               SnapshotTime Performance Positives Negatives TotalBins
#>     1: 2021-12-20 11:08:58    0.959459         1       148         2
#>     2: 2021-12-20 11:17:38    0.959459         1       148         2
#>     3: 2021-12-20 11:08:58    0.959459         1       148         2
#>     4: 2021-12-20 11:17:38    0.959459         1       148         2
#>     5: 2021-12-20 13:04:16    0.956044         1       273         2
#>    ---                                                              
#> 30367: 2021-12-20 13:04:16    0.520000       412      1953         4
#> 30368: 2021-12-20 12:32:50    0.520000       412      1953         4
#> 30369: 2021-12-20 13:04:16    0.520000       412      1953         4
#> 30370: 2021-12-20 12:32:50    0.520000       412      1953         4
#> 30371: 2021-12-20 13:04:16    0.520000       412      1953         4
#>        ResponseCount BinNegatives BinPositives BinResponseCount BinIndex
#>     1:           149          136            0              136        1
#>     2:           149          136            0              136        1
#>     3:           149           12            1               13        2
#>     4:           149           12            1               13        2
#>     5:           274          249            0              249        1
#>    ---                                                                  
#> 30367:          2365           78           26              104        2
#> 30368:          2365           35           12               47        3
#> 30369:          2365           35           12               47        3
#> 30370:          2365           47           12               59        4
#> 30371:          2365           47           12               59        4
#>             Lift     ZRatio GroupIndex FeatureImportance PredictorCategory
#>     1:  0.000000 -40.955260         29                 0                IH
#>     2:  0.000000 -40.955260         29                 0                IH
#>     3: 11.461538  40.955260         29                 0                IH
#>     4: 11.461538  40.955260         29                 0                IH
#>     5:  0.000000 -53.220062         31                 0                IH
#>    ---                                                                    
#> 30367:  1.435073   1.813905         60                 0                IH
#> 30368:  1.465606   1.271601         60                 0                IH
#> 30369:  1.465606   1.271601         60                 0                IH
#> 30370:  1.167517   0.563476         60                 0                IH
#> 30371:  1.167517   0.563476         60                 0                IH
#>        Propensity     AUC
#>     1: 0.00000000 95.9459
#>     2: 0.00000000 95.9459
#>     3: 0.07692308 95.9459
#>     4: 0.07692308 95.9459
#>     5: 0.00000000 95.6044
#>    ---                   
#> 30367: 0.25000000 52.0000
#> 30368: 0.25531915 52.0000
#> 30369: 0.25531915 52.0000
#> 30370: 0.20338983 52.0000
#> 30371: 0.20338983 52.0000