Subset the provided datamart data to just the latest snapshot per model.

If there is just one snapshot, nothing will change. It works for both model and predictor tables. If there is no snapshottime field, it will not do anything.

filterLatestSnapshotOnly(dt)

Arguments

dt

The data.table with the datamart data.

Value

A data.table with just the latest snapshots per model.

Examples

data(adm_datamart)
filterLatestSnapshotOnly(adm_datamart$modeldata)
#>                                  ModelID    Issue           Group
#>  1: 5ca51e1e-49f0-58d2-8ab7-4a0ccbb326fd    Sales DepositAccounts
#>  2: b216413c-8c2e-5e1b-860b-07dfc69efa69    Sales       AutoLoans
#>  3: cb67c091-5a81-56f4-9845-71afa9c8076e    Sales       HomeLoans
#>  4: ea0a225a-7d09-5d90-88f6-d0e8380d67cc    Sales DepositAccounts
#>  5: a8cde9d1-c720-537c-8fb8-8c59506bddfe    Sales     CreditCards
#>  6: cc9b9b1b-c2f9-5e8c-a54a-f13e0559ebd6    Sales     CreditCards
#>  7: 692f453a-f825-5a90-ab40-f83cbb34b058    Sales       HomeLoans
#>  8: 26dda9e0-34a2-5e0a-8ee2-fd0c39de729e    Sales     CreditCards
#>  9: f9fc995f-3f02-5e9e-8e22-b72909054f3a    Sales       AutoLoans
#> 10: 95ae0f10-297a-5c01-94d8-914bb89df969    Sales DepositAccounts
#> 11: 7ffc8c26-8473-5c8e-bcc3-48e5b162da7a    Sales     CreditCards
#> 12: 24f11deb-a6fa-59bf-b5f7-c61abffc9482    Sales       AutoLoans
#> 13: 84107ad6-c998-5377-b0d0-e02be2a8d768    Sales       AutoLoans
#> 14: b5338016-1fee-5745-80ee-7bd6d51f2920    Sales     CreditCards
#> 15: 9e9e8c2e-6d19-538c-8e7b-a73070717ecf    Sales     CreditCards
#> 16: c3f445b2-6037-5560-9f16-71a7259d2c2c    Sales       HomeLoans
#> 17: 59f06605-5b39-57dd-945b-637a72fa9085    Sales     CreditCards
#> 18: b27d8eb5-7ff6-5706-ae90-6d61ce695f78    Sales         Bundles
#> 19: c8a529fb-9d2c-5069-a50f-42ee1ea9341a Services         Account
#> 20: 602303a4-5243-523d-8964-473f3ea1200b    Sales     CreditCards
#> 21: 4efdb97d-288d-518f-a3e0-8314bbd7f2a3    Sales     CreditCards
#> 22: ab4592bc-df16-5372-86e7-0caea36ac877    Sales     CreditCards
#> 23: 8e40882a-aacf-5f03-9be8-f833c7932da2    Sales         Bundles
#> 24: a8cd5e12-5d10-5b62-95f6-c5bb6344a83a    Sales DepositAccounts
#> 25: 68e1d164-81e3-5da0-816c-0bbc3c20ac6c    Sales       AutoLoans
#> 26: 30cc2c18-5852-581a-abaa-555bf46a3b9c Services         Account
#> 27: a75db3a2-fecc-5cc8-b75f-0a44447f9cca    Sales       AutoLoans
#> 28: 98aeb30c-d896-51bd-b421-a650f0522335    Sales     CreditCards
#> 29: 34f7d1b9-92f7-57dc-ab62-0de35dda0221 Services         Account
#> 30: d5e2f4b1-249a-58c0-8a2f-2d314eada5fe    Sales       AutoLoans
#> 31: e4ea7a89-c394-5014-ac26-34b8eee475bd    Sales     CreditCards
#> 32: d6e132b6-ad92-5ece-8db5-b42d9d151f73    Sales         Bundles
#> 33: 08ca1302-9fc0-57bf-9031-d4179d400493    Sales         Bundles
#> 34: f3afaa9b-d0d5-5be9-a3b9-9fdfc4847411    Sales       AutoLoans
#> 35: 277a2c48-8888-5b71-911e-443d52c9b50f    Sales         Bundles
#> 36: 0ad9da35-f171-5ab9-973f-1cc4b89b4e8f    Sales         Bundles
#> 37: cffb1b5d-219e-5137-9036-4e2136be6a8f    Sales DepositAccounts
#> 38: 708f20ef-7ca1-58f4-9dfa-b7efb5342fcf    Sales       HomeLoans
#> 39: bd70a915-697a-5d43-ab2c-53b0557c85a0    Sales       AutoLoans
#> 40: 69a8e56f-6327-5a66-a9a3-9b0c106641ec    Sales       AutoLoans
#> 41: 708be173-0f46-55ee-add6-6fdca447a70a    Sales       HomeLoans
#> 42: 79e3f70f-dc6a-5bea-8006-02a6e91ad256    Sales DepositAccounts
#> 43: 320c4d01-43a5-510e-8fa8-1fc1bb4216da    Sales         Bundles
#> 44: c4cd4c00-87bf-562e-9ab8-20af64a81757    Sales DepositAccounts
#> 45: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7    Sales       HomeLoans
#> 46: c29ffcfa-5210-5998-a826-f3a8db57c610    Sales     CreditCards
#> 47: 0469930e-8d83-5c81-ba28-a8ec488f8e4d    Sales DepositAccounts
#> 48: 0b363d7f-68f8-5e9e-a843-40ca8849e05b    Sales       HomeLoans
#> 49: dd502275-e98b-51f7-8935-c30c84641871 Services         Account
#> 50: 2c8e753a-9e85-5060-ac5b-073e79e755c0    Sales     CreditCards
#> 51: 7d2046d5-be79-5f5a-a858-dcd9c65fd086    Sales       AutoLoans
#> 52: f38b5b50-b01b-5fd9-8120-7470320d6664    Sales DepositAccounts
#> 53: 081d0acf-bfc3-5c0a-abc4-29106252efb8    Sales     CreditCards
#> 54: a63f2f06-1e31-5691-a6df-48916a066c23    Sales DepositAccounts
#> 55: c4c1cff8-2879-51b8-8d08-cb961df94889    Sales       HomeLoans
#> 56: aaf75ad6-c71a-5fd8-b2eb-782cdbd30337 Services         Account
#> 57: baae77a0-099a-5104-b7a9-b9b1a23290fc Services        Customer
#> 58: 17571d69-f802-5122-9bfe-6f3f4ef9073b    Sales     CreditCards
#> 59: 7efc168f-9d8c-53ce-934f-c59d0c4bf314    Sales       HomeLoans
#> 60: 1be7cd72-741b-5d68-b148-c1af0142fd4a    Sales     CreditCards
#> 61: 838ca9ec-74ea-5a54-8c57-f45295116921 Services         Account
#> 62: a7bb008a-0514-5d41-b014-3e0b1b308382    Sales DepositAccounts
#> 63: 5cb8d642-390e-5c3e-8868-d34768f9535c    Sales       AutoLoans
#> 64: ebb6d905-7c43-5422-a813-f770feba7114    Sales       AutoLoans
#> 65: a87d51fb-cfdc-598d-b84a-dcdd64e1e750    Sales       HomeLoans
#> 66: 4954d7a5-b75e-5c71-89d6-e8340feb5022    Sales DepositAccounts
#> 67: 85d9dd89-93e5-55a0-b05f-e68bba9c97d1    Sales     CreditCards
#> 68: 1289c4af-1d3f-5265-9fe5-173e1ddaadc3    Sales       HomeLoans
#> 69: 3c892256-f415-5408-905d-3158345b10bf    Sales         Bundles
#> 70: bb02128b-d6df-52f3-9983-316f9fce47c7    Sales     CreditCards
#>                                  ModelID    Issue           Group
#>                                Name Channel Direction Treatment
#>  1:                   RegularSaving     SMS  Outbound        NA
#>  2:                AutoUsed36Months     SMS  Outbound        NA
#>  3:                FirstMortgageFHA   Email  Outbound        NA
#>  4:                 PremierChecking     Web   Inbound        NA
#>  5:                    UPlusFinGold     Web   Inbound        NA
#>  6:         IdentityTheftProtection   Email  Outbound        NA
#>  7:              FirstMortgageFloat     Web   Inbound        NA
#>  8:          ChannelAction_Template     SMS  Outbound        NA
#>  9:                 AutoNew36Months     SMS  Outbound        NA
#> 10:       MoneyMarketSavingsAccount     Web   Inbound        NA
#> 11:                 MasterCardWorld     SMS  Outbound        NA
#> 12:                AutoUsed84Months   Email  Outbound        NA
#> 13:                AutoUsed60Months   Email  Outbound        NA
#> 14:               PaymentProtection   Email  Outbound        NA
#> 15:                  MasterCardGold     Web   Inbound        NA
#> 16:                   FirstMortgage     SMS  Outbound        NA
#> 17:             PlatinumRewardsCard     Web   Inbound        NA
#> 18:                  PremiumBanking     SMS  Outbound        NA
#> 19:          IncreaseYourCreditLine     Web   Inbound        NA
#> 20:                UPlusFinPersonal     SMS  Outbound        NA
#> 21:                  MasterCardGold     SMS  Outbound        NA
#> 22:                UPlusFinPersonal   Email  Outbound        NA
#> 23:                  PremiumBanking     Web   Inbound        NA
#> 24:                  SeniorChecking     Web   Inbound        NA
#> 25:                 AutoNew84Months     Web   Inbound        NA
#> 26: Services_ChannelAction_Template   Email  Outbound        NA
#> 27:                AutoUsed48Months     SMS  Outbound        NA
#> 28:         IdentityTheftProtection     SMS  Outbound        NA
#> 29: Services_ChannelAction_Template     SMS  Outbound        NA
#> 30:                AutoUsed84Months     SMS  Outbound        NA
#> 31:               PaymentProtection     SMS  Outbound        NA
#> 32:                      SuperSaver     SMS  Outbound        NA
#> 33:                      HomeOwners     Web   Inbound        NA
#> 34:                 AutoNew84Months   Email  Outbound        NA
#> 35:                      HomeOwners     SMS  Outbound        NA
#> 36:                      SuperSaver     Web   Inbound        NA
#> 37:                 PremierChecking   Email  Outbound        NA
#> 38:         FirstMortgageFiveOneARM     Web   Inbound        NA
#> 39:                 AutoNew36Months   Email  Outbound        NA
#> 40:                 AutoNew48Months     SMS  Outbound        NA
#> 41:         FirstMortgageFiveOneARM     SMS  Outbound        NA
#> 42:       MoneyMarketSavingsAccount   Email  Outbound        NA
#> 43:                  PremiumBanking   Email  Outbound        NA
#> 44:                   RegularSaving   Email  Outbound        NA
#> 45:              FirstMortgageFloat   Email  Outbound        NA
#> 46:               PaymentProtection     Web   Inbound        NA
#> 47:                  SeniorChecking     SMS  Outbound        NA
#> 48:         FirstMortgageFiveOneARM   Email  Outbound        NA
#> 49: Services_ChannelAction_Template     Web   Inbound        NA
#> 50:                    UPlusFinGold   Email  Outbound        NA
#> 51:                 AutoNew36Months     Web   Inbound        NA
#> 52:       MoneyMarketSavingsAccount     SMS  Outbound        NA
#> 53:                    UPlusFinGold     SMS  Outbound        NA
#> 54:                   BasicChecking   Email  Outbound        NA
#> 55:        FirstMortgageThreeOneARM   Email  Outbound        NA
#> 56:          IncreaseYourCreditLine     SMS  Outbound        NA
#> 57:       MissingContactInformation     SMS  Outbound        NA
#> 58:                        VisaGold     Web   Inbound        NA
#> 59:        FirstMortgageSevenOneARM   Email  Outbound        NA
#> 60:                        VisaGold     SMS  Outbound        NA
#> 61:                     GoPaperless     SMS  Outbound        NA
#> 62:                 PremierChecking     SMS  Outbound        NA
#> 63:                AutoUsed36Months     Web   Inbound        NA
#> 64:                 AutoNew48Months   Email  Outbound        NA
#> 65:                   FirstMortgage     Web   Inbound        NA
#> 66:                  SeniorChecking   Email  Outbound        NA
#> 67:                  MasterCardGold   Email  Outbound        NA
#> 68:                FirstMortgageFHA     SMS  Outbound        NA
#> 69:                   StudentChoice   Email  Outbound        NA
#> 70:             PlatinumRewardsCard   Email  Outbound        NA
#>                                Name Channel Direction Treatment
#>     ConfigurationName ActivePredictors TotalPredictors Negatives Positives
#>  1: OmniAdaptiveModel               36              89      1143        78
#>  2: OmniAdaptiveModel               30              89      1480       142
#>  3: OmniAdaptiveModel               15              89       349        13
#>  4: OmniAdaptiveModel               21              89      3181      1223
#>  5: OmniAdaptiveModel               38              89      3537       223
#>  6: OmniAdaptiveModel               14              89       806         2
#>  7: OmniAdaptiveModel               10              89       505        13
#>  8: OmniAdaptiveModel               34              89       776        52
#>  9: OmniAdaptiveModel               24              89      5566        36
#> 10: OmniAdaptiveModel               17              89      1096       110
#> 11: OmniAdaptiveModel               27              89       598        36
#> 12: OmniAdaptiveModel               13              89      1230        13
#> 13: OmniAdaptiveModel               36              89      1324        54
#> 14: OmniAdaptiveModel               24              89       508        29
#> 15: OmniAdaptiveModel               11              89       817        18
#> 16: OmniAdaptiveModel               36              89      3281       665
#> 17: OmniAdaptiveModel               17              89       893        14
#> 18: OmniAdaptiveModel               26              89      4949       103
#> 19: OmniAdaptiveModel               34              89      6200       271
#> 20: OmniAdaptiveModel               35              89      3402       186
#> 21: OmniAdaptiveModel               12              89      1077         3
#> 22: OmniAdaptiveModel               15              89      3454        54
#> 23: OmniAdaptiveModel               29              89      5075       101
#> 24: OmniAdaptiveModel               28              89      2449       575
#> 25: OmniAdaptiveModel               36              89      1430       206
#> 26: OmniAdaptiveModel               32              89      7361        54
#> 27: OmniAdaptiveModel               31              89      1396        51
#> 28: OmniAdaptiveModel               12              89       845        10
#> 29: OmniAdaptiveModel               36              89      7050       237
#> 30: OmniAdaptiveModel               35              89      1297        50
#> 31: OmniAdaptiveModel               34              89       560       162
#> 32: OmniAdaptiveModel               31              89       973       133
#> 33: OmniAdaptiveModel               33              89      3996       609
#> 34: OmniAdaptiveModel               33              89      1117        67
#> 35: OmniAdaptiveModel               33              89      4440       536
#> 36: OmniAdaptiveModel               34              89      1155       112
#> 37: OmniAdaptiveModel               29              89      3063       292
#> 38: OmniAdaptiveModel               33              89       371        78
#> 39: OmniAdaptiveModel                9              89      5805        12
#> 40: OmniAdaptiveModel               31              89      5330       413
#> 41: OmniAdaptiveModel               31              89       432        50
#> 42: OmniAdaptiveModel               10              89      1077        19
#> 43: OmniAdaptiveModel               27              89      5406        24
#> 44: OmniAdaptiveModel               14              89      1054        15
#> 45: OmniAdaptiveModel               13              89       442         2
#> 46: OmniAdaptiveModel               39              89       514       172
#> 47: OmniAdaptiveModel               31              89      2882       542
#> 48: OmniAdaptiveModel               11              89       394        10
#> 49: OmniAdaptiveModel               33              89      6713       360
#> 50: OmniAdaptiveModel               31              89      3507        38
#> 51: OmniAdaptiveModel               21              89      5362       137
#> 52: OmniAdaptiveModel               33              89      1067       118
#> 53: OmniAdaptiveModel               37              89      3689       165
#> 54: OmniAdaptiveModel               28              89      3586       282
#> 55: OmniAdaptiveModel               12              89       537         2
#> 56: OmniAdaptiveModel               37              89      7468       200
#> 57: OmniAdaptiveModel               28              89      5752      1344
#> 58: OmniAdaptiveModel               28              89       732        46
#> 59: OmniAdaptiveModel                7              89       300         5
#> 60: OmniAdaptiveModel               23              89       795        27
#> 61: OmniAdaptiveModel               35              89      7270       904
#> 62: OmniAdaptiveModel               16              89      3556      1102
#> 63: OmniAdaptiveModel               35              89      1515       173
#> 64: OmniAdaptiveModel               32              89      5055       143
#> 65: OmniAdaptiveModel               30              89      3173       903
#> 66: OmniAdaptiveModel               36              89      2737       146
#> 67: OmniAdaptiveModel               13              89       870         5
#> 68: OmniAdaptiveModel               25              89       429        48
#> 69: OmniAdaptiveModel               29              89      5115        88
#> 70: OmniAdaptiveModel               13              89       987         2
#>     ConfigurationName ActivePredictors TotalPredictors Negatives Positives
#>     ResponseCount        SnapshotTime SuccessRate Performance Extension     AUC
#>  1:          1221 2021-12-20 14:57:49 0.063882064    0.511978        NA 51.1978
#>  2:          1622 2021-12-20 14:57:49 0.087546239    0.542851        NA 54.2851
#>  3:           362 2021-12-20 14:57:49 0.035911602    0.676805        NA 67.6805
#>  4:          4404 2021-12-20 14:57:49 0.277702089    0.524186        NA 52.4186
#>  5:          3760 2021-12-20 14:57:49 0.059308511    0.686745        NA 68.6745
#>  6:           808 2021-12-20 14:57:49 0.002475248    0.500000        NA 50.0000
#>  7:           518 2021-12-20 14:57:49 0.025096525    0.646733        NA 64.6733
#>  8:           828 2021-12-20 14:57:49 0.062801932    0.697884        NA 69.7884
#>  9:          5602 2021-12-20 14:57:49 0.006426276    0.605845        NA 60.5845
#> 10:          1206 2021-12-20 14:57:49 0.091210614    0.500000        NA 50.0000
#> 11:           634 2021-12-20 14:57:49 0.056782334    0.551591        NA 55.1591
#> 12:          1243 2021-12-20 14:57:49 0.010458568    0.500000        NA 50.0000
#> 13:          1378 2021-12-20 14:57:49 0.039187228    0.653775        NA 65.3775
#> 14:           537 2021-12-20 14:57:49 0.054003724    0.525041        NA 52.5041
#> 15:           835 2021-12-20 14:57:49 0.021556886    0.571711        NA 57.1711
#> 16:          3946 2021-12-20 14:57:49 0.168525089    0.648510        NA 64.8510
#> 17:           907 2021-12-20 14:57:49 0.015435502    0.579772        NA 57.9772
#> 18:          5052 2021-12-20 14:57:49 0.020387965    0.616472        NA 61.6472
#> 19:          6471 2021-12-20 14:57:49 0.041879153    0.592788        NA 59.2788
#> 20:          3588 2021-12-20 14:57:49 0.051839465    0.629893        NA 62.9893
#> 21:          1080 2021-12-20 14:57:49 0.002777778    0.500000        NA 50.0000
#> 22:          3508 2021-12-20 14:57:49 0.015393387    0.619665        NA 61.9665
#> 23:          5176 2021-12-20 14:57:49 0.019513138    0.652640        NA 65.2640
#> 24:          3024 2021-12-20 14:57:49 0.190145503    0.549147        NA 54.9147
#> 25:          1636 2021-12-20 14:57:49 0.125916870    0.774901        NA 77.4901
#> 26:          7415 2021-12-20 14:57:49 0.007282535    0.629442        NA 62.9442
#> 27:          1447 2021-12-20 14:57:49 0.035245335    0.693922        NA 69.3922
#> 28:           855 2021-12-20 14:57:49 0.011695906    0.530707        NA 53.0707
#> 29:          7287 2021-12-20 14:57:49 0.032523672    0.710618        NA 71.0618
#> 30:          1347 2021-12-20 14:57:49 0.037119525    0.622813        NA 62.2813
#> 31:           722 2021-12-20 14:57:49 0.224376731    0.514945        NA 51.4945
#> 32:          1106 2021-12-20 14:57:49 0.120253165    0.618120        NA 61.8120
#> 33:          4605 2021-12-20 14:57:49 0.132247557    0.756002        NA 75.6002
#> 34:          1184 2021-12-20 14:57:49 0.056587838    0.642461        NA 64.2461
#> 35:          4976 2021-12-20 14:57:49 0.107717042    0.773211        NA 77.3211
#> 36:          1267 2021-12-20 14:57:49 0.088397790    0.757296        NA 75.7296
#> 37:          3355 2021-12-20 14:57:49 0.087034277    0.521103        NA 52.1103
#> 38:           449 2021-12-20 14:57:49 0.173719376    0.527198        NA 52.7198
#> 39:          5817 2021-12-20 14:57:49 0.002062919    0.523734        NA 52.3734
#> 40:          5743 2021-12-20 14:57:49 0.071913634    0.618246        NA 61.8246
#> 41:           482 2021-12-20 14:57:49 0.103734440    0.652020        NA 65.2020
#> 42:          1096 2021-12-20 14:57:49 0.017335766    0.500000        NA 50.0000
#> 43:          5430 2021-12-20 14:57:49 0.004419890    0.624670        NA 62.4670
#> 44:          1069 2021-12-20 14:57:49 0.014031805    0.614512        NA 61.4512
#> 45:           444 2021-12-20 14:57:49 0.004504505    0.500000        NA 50.0000
#> 46:           686 2021-12-20 14:57:49 0.250728863    0.581586        NA 58.1586
#> 47:          3424 2021-12-20 14:57:49 0.158294393    0.565227        NA 56.5227
#> 48:           404 2021-12-20 14:57:49 0.024752475    0.530007        NA 53.0007
#> 49:          7073 2021-12-20 14:57:49 0.050897780    0.663954        NA 66.3954
#> 50:          3545 2021-12-20 14:57:49 0.010719323    0.534960        NA 53.4960
#> 51:          5499 2021-12-20 14:57:49 0.024913621    0.565041        NA 56.5041
#> 52:          1185 2021-12-20 14:57:49 0.099578059    0.513569        NA 51.3569
#> 53:          3854 2021-12-20 14:57:49 0.042812662    0.665410        NA 66.5410
#> 54:          3868 2021-12-20 14:57:49 0.072905895    0.510276        NA 51.0276
#> 55:           539 2021-12-20 14:52:44 0.003710575    0.500000        NA 50.0000
#> 56:          7668 2021-12-20 14:57:49 0.026082420    0.564527        NA 56.4527
#> 57:          7096 2021-12-20 14:57:49 0.189402480    0.622187        NA 62.2187
#> 58:           778 2021-12-20 14:57:49 0.059125964    0.519586        NA 51.9586
#> 59:           305 2021-12-20 14:19:43 0.016393443    0.566071        NA 56.6071
#> 60:           822 2021-12-20 14:57:49 0.032846715    0.531863        NA 53.1863
#> 61:          8174 2021-12-20 14:57:49 0.110594568    0.712165        NA 71.2165
#> 62:          4658 2021-12-20 14:57:49 0.236582224    0.516802        NA 51.6802
#> 63:          1688 2021-12-20 14:57:49 0.102488152    0.578638        NA 57.8638
#> 64:          5198 2021-12-20 14:57:49 0.027510581    0.580125        NA 58.0125
#> 65:          4076 2021-12-20 14:57:49 0.221540726    0.647708        NA 64.7708
#> 66:          2883 2021-12-20 14:57:49 0.050641693    0.549252        NA 54.9252
#> 67:           875 2021-12-20 14:57:49 0.005714286    0.500000        NA 50.0000
#> 68:           477 2021-12-20 14:57:49 0.100628931    0.527779        NA 52.7779
#> 69:          5203 2021-12-20 14:57:49 0.016913319    0.601887        NA 60.1887
#> 70:           989 2021-12-20 14:57:49 0.002022245    0.500000        NA 50.0000
#>     ResponseCount        SnapshotTime SuccessRate Performance Extension     AUC