Remove the binning information from the provided datamart predictor data so it only is predictor level, not more detailed. This is commonly used in plotting functions.

Remove the binning information from the provided datamart predictor data so it only is predictor level, not more detailed. This is commonly used in plotting functions.

filterPredictorBinning(dt)

Arguments

dt

The data.table with the datamart predictor data.

Value

A data.table without the predictor binning.

Examples

data(adm_datamart)
filterPredictorBinning(adm_datamart$predictordata)
#>                                              PredictorName
#>     1:               IH.SMS.Outbound.Churned.pxLastGroupID
#>     2:               IH.SMS.Outbound.Churned.pxLastGroupID
#>     3: IH.SMS.Outbound.Churned.pxLastOutcomeTime.DaysSince
#>     4: IH.SMS.Outbound.Churned.pxLastOutcomeTime.DaysSince
#>     5:    IH.SMS.Outbound.Churned.pyHistoricalOutcomeCount
#>    ---                                                    
#> 32393:                          Param.ExtGroupWealthoffers
#> 32394:                          Param.ExtGroupWealthoffers
#> 32395:                          Param.ExtGroupWealthoffers
#> 32396:                          Param.ExtGroupWealthoffers
#> 32397:                          Param.ExtGroupWealthoffers
#>                                     ModelID     Type      BinType EntryType
#>     1: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7 symbolic      MISSING    Active
#>     2: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7 symbolic      MISSING    Active
#>     3: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7  numeric      MISSING  Inactive
#>     4: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7  numeric      MISSING  Inactive
#>     5: 470485f7-3b20-58f7-a3aa-1f12c2bb99b7  numeric      MISSING  Inactive
#>    ---                                                                     
#> 32393: 277a2c48-8888-5b71-911e-443d52c9b50f symbolic EQUIBEHAVIOR  Inactive
#> 32394: 277a2c48-8888-5b71-911e-443d52c9b50f symbolic EQUIBEHAVIOR  Inactive
#> 32395: 277a2c48-8888-5b71-911e-443d52c9b50f symbolic EQUIBEHAVIOR  Inactive
#> 32396: 277a2c48-8888-5b71-911e-443d52c9b50f symbolic EQUIBEHAVIOR  Inactive
#> 32397: 277a2c48-8888-5b71-911e-443d52c9b50f symbolic EQUIBEHAVIOR  Inactive
#>               SnapshotTime Performance Positives Negatives TotalBins
#>     1: 2021-12-20 11:08:58    0.959459         1       148         2
#>     2: 2021-12-20 11:17:38    0.959459         1       148         2
#>     3: 2021-12-20 11:08:58    0.959459         1       148         2
#>     4: 2021-12-20 11:17:38    0.959459         1       148         2
#>     5: 2021-12-20 11:08:58    0.959459         1       148         2
#>    ---                                                              
#> 32393: 2021-12-20 14:19:46    0.500000       484      3988         1
#> 32394: 2021-12-20 14:57:49    0.500000       536      4440         1
#> 32395: 2021-12-20 13:45:59    0.500000       370      3104         1
#> 32396: 2021-12-20 13:49:28    0.500000       429      3551         1
#> 32397: 2021-12-20 14:52:45    0.500000       484      3988         1
#>        ResponseCount GroupIndex FeatureImportance PredictorCategory Propensity
#>     1:           149         29                 0                IH  0.0000000
#>     2:           149         29                 0                IH  0.0000000
#>     3:           149         29                 0                IH  0.0000000
#>     4:           149         29                 0                IH  0.0000000
#>     5:           149         29                 0                IH  0.0000000
#>    ---                                                                        
#> 32393:          4472          2                 0             Param  0.1082290
#> 32394:          4976          2                 0             Param  0.1077170
#> 32395:          3474          2                 0             Param  0.1065055
#> 32396:          3980          2                 0             Param  0.1077889
#> 32397:          4472          2                 0             Param  0.1082290
#>            AUC
#>     1: 95.9459
#>     2: 95.9459
#>     3: 95.9459
#>     4: 95.9459
#>     5: 95.9459
#>    ---        
#> 32393: 50.0000
#> 32394: 50.0000
#> 32395: 50.0000
#> 32396: 50.0000
#> 32397: 50.0000